requirements.txt
promptflow
promptflow-tools
promptflow-core
nltk
transformers
torch
scikit-learn
numpy
python-dotenv
openai
python-docx
fastapi
uvicorn